UserDao misc and test

+++ conf/evolutions/default/1.sql
@@ -19,7 +19,7 @@ create table feed_articles (
 );
 
 create table users (
-  id                            bigint auto_increment not null,
+  id                            integer auto_increment not null,
   username                      varchar(255),
   password                      varchar(255),
   mail                          varchar(255),
@@ -30,7 +30,7 @@ create table users (
 
 create table user_articles (
   id                            integer auto_increment not null,
-  user_id                       bigint,
+  user_id                       integer,
   feed_articles_id              integer,
   is_read                       tinyint(1) default 0,
   constraint pk_user_articles primary key (id)
@@ -38,7 +38,7 @@ create table user_articles (
 
 create table user_subscriptions (
   id                            integer auto_increment not null,
-  user_id                       bigint,
+  user_id                       integer,
   feed_id                       integer,
   constraint pk_user_subscriptions primary key (id)
 );
